renamed testserver
